Re-use screenshot code
